Create finalProjectRCode
It's just the R code used for the paper I submitted. There's a LOT more scrap code left behind in a .txt file on my computer but this is what the final end result is, where if you copy and paste it into R it'll spit out all the tests and plots I used.